+++ module/skills.lua
@@ -16,8 +16,9 @@ function ns:updateSkills ()
 		end
 	end
 
-	if (hasFilter == false) then
+	if (numSkills > 1 and hasFilter == false) then
 		local header = 'undefined';
+		local names = {};
 		Pulse.realm[realm].char[char].skills = {};
 		for i = 1, numSkills do
 			local skillName, isHeader, isExpanded, skillRank, numTempPoints, skillModifier, skillMaxRank, isAbandonable, stepCost, rankCost, minLevel, skillCostType, skillDescription = GetSkillLineInfo(i);
@@ -27,6 +28,7 @@ function ns:updateSkills ()
 			else
 				local temp = {};
 				temp.skillName = skillName;
+				tinsert(names, skillName);
 				temp.skillRank = skillRank;
 				temp.numTempPoints = numTempPoints;
 				temp.skillModifier = skillModifier;
@@ -43,6 +45,14 @@ function ns:updateSkills ()
 				end
 
 				tinsert(Pulse.realm[realm].char[char].skills[header], temp);
+			end
+		end
+
+		if (Pulse.realm[realm].char[char].tradeskills ~= nil) then
+			for k in pairs(Pulse.realm[realm].char[char].tradeskills) do
+				if (not tContains(names, k)) then
+					Pulse.realm[realm].char[char].tradeskills[k] = nil;
+				end
 			end
 		end
 	end