+++ module/talents.lua
@@ -9,6 +9,10 @@ function ns:updateTalents ()
 	local numTabs = GetNumTalentTabs();
 
 	for i = 1, numTabs do
+		local talentName = GetTalentTabInfo(i);
+
+		Pulse.realm[realm].char[char].talents[talentName] = {};
+
 		local numTalents = GetNumTalents(i);
 		for j = 1, numTalents do
 			local name, icon, tier, column, rank, maxRank = GetTalentInfo(i, j);
@@ -21,7 +25,7 @@ function ns:updateTalents ()
 			temp.rank = rank;
 			temp.maxRank = maxRank;
 
-			tinsert(Pulse.realm[realm].char[char].talents, temp);
+			tinsert(Pulse.realm[realm].char[char].talents[talentName], temp);
 		end
 	end
 end